# openapi.yaml
openapi: '3.0.3'
info:
  title: Career-Copilot API
  description: AI模拟面试官与智能职业规划平台 - 后端接口文档
  version: '1.0.0'
  contact:
    name: 陶宏阳
    url: https://github.com/peeker-tao
servers:
  - url: http://localhost:3002/api
    description: 本地开发环境
  - url: https://test-api.career-copilot.com/api/v1
    description: 测试环境
  - url: https://api.career-copilot.com/api/v1
    description: 生产环境
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 在 Authorization 头中传入 Bearer Token
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: HTTP 状态码
          example: 200
        message:
          type: string
          description: 提示信息
          example: success
        data:
          type: object
          description: 业务数据
      required:
        - code
        - message
        - data
...
    PaginationResult:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
...
    ResumeSummary:
      type: object
      properties:
        id:
          type: string
          example: clx...
        title:
          type: string
          example: 小明_后端简历.pdf
        status:
          type: string
          enum:
            - parsing
            - completed
            - failed
          example: completed
        skills:
          type: array
          items:
            type: string
          example:
            - Java
            - Spring Boot
            - MySQL
            - Redis
        createdAt:
          type: string
          format: date-time
          example: '2026-06-13T08:00:00.000Z'
    Education:
      type: object
      properties:
        school:
          type: string
          example: 华中科技大学
        major:
          type: string
          example: 软件工程
        degree:
          type: string
          example: 本科
        period:
          type: string
          example: '2022-2026'
    Experience:
      type: object
      properties:
        company:
          type: string
          example: XX科技
        position:
          type: string
          example: 后端开发实习生
        period:
          type: string
          example: '2025.06-2025.09'
        description:
          type: string
          example: 负责后端 API 开发与维护
    Project:
      type: object
      properties:
        name:
          type: string
          example: 在线商城系统
        role:
          type: string
          example: 后端开发
        techStack:
          type: array
          items:
            type: string
          example:
            - Spring Boot
            - MySQL
        description:
          type: string
          example: 基于 Spring Boot 的电商后端系统
    ParsedResumeData:
      type: object
      properties:
        basicInfo:
          type: object
          properties:
            name:
              type: string
              example: 小明
            phone:
              type: string
              example: 138****1234
            email:
              type: string
              format: email
              example: user@example.com
        education:
          type: array
          items:
            $ref: '#/components/schemas/Education'
        experience:
          type: array
          items:
            $ref: '#/components/schemas/Experience'
        skills:
          type: array
          items:
            type: string
          example:
            - Java
            - Spring Boot
            - MySQL
            - Redis
            - Git
            - Docker
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
    ResumeDetail:
      allOf:
        - $ref: '#/components/schemas/ResumeSummary'
        - type: object
          properties:
            fileUrl:
              type: string
              format: uri
              nullable: true
              example: https://...
            parsedData:
              $ref: '#/components/schemas/ParsedResumeData'
    UpdateResumeRequest:
      type: object
      properties:
        parsedData:
          type: object
          properties:
            skills:
              type: array
              items:
                type: string
              example:
                - Java
                - Spring Boot
                - MySQL
                - Redis
                - RabbitMQ

    # ============================================================
    # 面试模块
    # ============================================================
...
paths:
...
  /resumes/upload:
    post:
      tags:
        - 简历管理
      summary: 上传简历
      description: 上传 PDF 或 Word 格式的简历文件，系统将异步解析
      operationId: uploadResume
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: 简历文件，支持 .pdf / .docx，≤ 10MB
                title:
                  type: string
                  description: 简历名称（默认使用文件名）
                  example: 小明_简历_2026
      responses:
        '201':
          description: 上传成功，简历异步解析中
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      code:
                        type: integer
                        example: 201
                      message:
                        type: string
                        example: 上传成功，简历解析中
                      data:
                        type: object
                        properties:
                          id:
                            type: string
                            example: clx...
                          title:
                            type: string
                            example: 小明_简历_2026.pdf
                          fileUrl:
                            type: string
                            format: uri
                            example: https://...
                          status:
                            type: string
                            enum:
                              - parsing
                            example: parsing
        '400':
          description: 参数错误
        '413':
          description: 文件大小超限（> 10MB）
  /resumes:
    get:
      tags:
        - 简历管理
      summary: 获取简历列表
      description: 获取当前用户的所有简历列表
      operationId: listResumes
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: 页码
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
          description: 每页条数
      responses:
        '200':
          description: 成功获取简历列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/ResumeSummary'
                          pagination:
                            $ref: '#/components/schemas/PaginationResult'
  /resumes/{id}:
    get:
      tags:
        - 简历管理
      summary: 获取简历详情
      description: 获取单份简历的详细信息，包含 LLM 解析后的结构化数据
      operationId: getResume
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 简历 ID
          example: clx...
      responses:
        '200':
          description: 成功获取简历详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ResumeDetail'
        '404':
          description: 简历不存在
    put:
      tags:
        - 简历管理
      summary: 编辑简历解析结果
      description: 手动修正 LLM 解析后的简历数据
      operationId: updateResume
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 简历 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResumeRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ResumeDetail'
        '404':
          description: 简历不存在
    delete:
      tags:
        - 简历管理
      summary: 删除简历
      description: 删除指定简历
      operationId: deleteResume
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 简历 ID
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      message:
                        type: string
                        example: 删除成功
        '404':
          description: 简历不存在

  # ============================================================
  # 面试模块
  # ============================================================
...